refactor(admin): use non index file for components
