+++ admin/src/components/NoteListLayout/components/NoteListLayoutHeader/NoteListLayoutHeader.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Box, Divider, Typography } from '@strapi/design-system';
+
+const NoteListLayoutHeader = () => {
+	return (
+		<React.Fragment>
+			<Typography variant="sigma" textColor="neutral600">
+				Notes
+			</Typography>
+			<Box marginTop={2}>
+				<Divider />
+			</Box>
+		</React.Fragment>
+	);
+};
+
+export default NoteListLayoutHeader;
