+++ admin/src/components/NoteListLayout/components/NoteListLayoutHeader/index.js
@@ -1,17 +1,1 @@
-import React from 'react';
-import { Box, Divider, Typography } from '@strapi/design-system';
-
-const NoteListLayoutHeader = () => {
-	return (
-		<React.Fragment>
-			<Typography variant="sigma" textColor="neutral600">
-				Notes
-			</Typography>
-			<Box marginTop={2}>
-				<Divider />
-			</Box>
-		</React.Fragment>
-	);
-};
-
-export { NoteListLayoutHeader };
+export { default } from './NoteListLayoutHeader';
